lemma_rs_be/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.1.4
django-cors-headers==3.7.0
django-oauth-toolkit==1.3.3
djangorestframework==3.12.2
drf-social-oauth2==1.0.8
drf-writable-nested==0.6.2
drf-yasg==1.20.0
ecdsa==0.14.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.7
python-jose==3.2.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
